+++ CHANGELOG.md
@@ -2,8 +2,24 @@
 
 ## Unreleased
 
-- Migrated to Rust 2021.
+### Changed
+
+- Textual representation: 26-digit Base32 -> 25-digit Base36
+- Field structure: { `timestamp`: 44 bits, `counter`: 28 bits, `per_sec_random`:
+  24 bits, `per_gen_random`: 32 bits } -> { `timestamp`: 48 bits, `counter_hi`:
+  24 bits, `counter_lo`: 24 bits, `entropy`: 32 bits }
+- Timestamp epoch: 2020-01-01 00:00:00.000 UTC -> 1970-01-01 00:00:00.000 UTC
+- Rust edition: 2018 -> 2021
+
+### Added
+
+- `Scru128Id#counter_hi()`, `Scru128Id#counter_lo()`, `Scru128Id#entropy()`
+
+### Removed
+
+- `TIMESTAMP_BIAS`
+- `Scru128Id#counter()`, `Scru128Id#per_sec_random()`, `Scru128Id#per_gen_random()`
 
 ## v1.0.0 - 2022-01-03
 
-- Initial stable release.
+- Initial stable release